Regroup all the formatting function in one module
