+++ charon/src/ast/expressions_utils.rs
@@ -1,6 +1,5 @@
 //! This file groups everything which is linked to implementations about [crate::expressions]
 use crate::expressions::*;
-use crate::formatter::{AstFormatter, FmtCtx};
 use crate::gast::{AssumedFunId, Call, FnOperand, FunId, FunIdOrTraitMethodRef, TraitItemName};
 use crate::types::*;
 use crate::ullbc_ast::GlobalDeclId;
@@ -14,298 +13,6 @@ impl Place {
             var_id,
             projection: Vec::new(),
         }
-    }
-}
-
-impl std::fmt::Display for BorrowKind {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::result::Result<(), std::fmt::Error> {
-        match self {
-            BorrowKind::Shared => write!(f, "Shared"),
-            BorrowKind::Mut => write!(f, "Mut"),
-            BorrowKind::TwoPhaseMut => write!(f, "TwoPhaseMut"),
-            BorrowKind::Shallow => write!(f, "Shallow"),
-        }
-    }
-}
-
-impl CastKind {
-    pub fn fmt_with_ctx<C>(&self, ctx: &C) -> String
-    where
-        C: AstFormatter,
-    {
-        match self {
-            CastKind::Scalar(src, tgt) => format!("cast<{src},{tgt}>"),
-            CastKind::FnPtr(src, tgt) => {
-                format!("cast<{},{}>", src.fmt_with_ctx(ctx), tgt.fmt_with_ctx(ctx))
-            }
-        }
-    }
-}
-
-impl UnOp {
-    pub fn fmt_with_ctx<C>(&self, ctx: &C) -> String
-    where
-        C: AstFormatter,
-    {
-        match self {
-            UnOp::Not => "~".to_string(),
-            UnOp::Neg => "-".to_string(),
-            UnOp::Cast(kind) => kind.fmt_with_ctx(ctx),
-            UnOp::ArrayToSlice(..) => "array_to_slice".to_string(),
-        }
-    }
-}
-
-impl std::fmt::Display for BinOp {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::result::Result<(), std::fmt::Error> {
-        match self {
-            BinOp::BitXor => write!(f, "^"),
-            BinOp::BitAnd => write!(f, "&"),
-            BinOp::BitOr => write!(f, "|"),
-            BinOp::Eq => write!(f, "=="),
-            BinOp::Lt => write!(f, "<"),
-            BinOp::Le => write!(f, "<="),
-            BinOp::Ne => write!(f, "!="),
-            BinOp::Ge => write!(f, ">="),
-            BinOp::Gt => write!(f, ">"),
-            BinOp::Div => write!(f, "/"),
-            BinOp::Rem => write!(f, "%"),
-            BinOp::Add => write!(f, "+"),
-            BinOp::Sub => write!(f, "-"),
-            BinOp::Mul => write!(f, "*"),
-            BinOp::Shl => write!(f, "<<"),
-            BinOp::Shr => write!(f, ">>"),
-        }
-    }
-}
-
-impl Place {
-    pub fn fmt_with_ctx<C>(&self, ctx: &C) -> String
-    where
-        C: AstFormatter,
-    {
-        let mut out = ctx.format_object(self.var_id);
-
-        for p in &self.projection {
-            match p {
-                ProjectionElem::Deref => {
-                    out = format!("*({out})");
-                }
-                ProjectionElem::DerefBox => {
-                    out = format!("deref_box ({out})");
-                }
-                ProjectionElem::DerefRawPtr => {
-                    out = format!("deref_raw_ptr ({out})");
-                }
-                ProjectionElem::Field(proj_kind, field_id) => match proj_kind {
-                    FieldProjKind::Adt(adt_id, opt_variant_id) => {
-                        let field_name = ctx.format_object((*adt_id, *opt_variant_id, *field_id));
-                        let downcast = match opt_variant_id {
-                            None => "".to_string(),
-                            Some(variant_id) => format!(" as variant @{variant_id}"),
-                        };
-                        out = format!("({out}{downcast}).{field_name}");
-                    }
-                    FieldProjKind::Tuple(_) => {
-                        out = format!("({out}).{field_id}");
-                    }
-                    FieldProjKind::ClosureState => {
-                        out = format!("({out}).@closure_state_field_{field_id}");
-                    }
-                },
-                ProjectionElem::Index(i, _) => out = format!("({out})[{}]", ctx.format_object(*i)),
-            }
-        }
-
-        out
-    }
-}
-
-impl std::fmt::Display for Place {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::result::Result<(), std::fmt::Error> {
-        write!(f, "{}", self.fmt_with_ctx(&FmtCtx::new()))
-    }
-}
-
-impl ConstantExpr {
-    pub fn fmt_with_ctx<C>(&self, ctx: &C) -> String
-    where
-        C: AstFormatter,
-    {
-        self.value.fmt_with_ctx(ctx)
-    }
-}
-
-impl RawConstantExpr {
-    pub fn fmt_with_ctx<C>(&self, ctx: &C) -> String
-    where
-        C: AstFormatter,
-    {
-        match self {
-            RawConstantExpr::Literal(c) => c.to_string(),
-            RawConstantExpr::Adt(variant_id, values) => {
-                // It is a bit annoying: in order to properly format the value,
-                // we need the type (which contains the type def id).
-                // Anyway, the printing utilities are mostly for debugging.
-                let variant_id = match variant_id {
-                    Option::Some(id) => format!("Some({id})"),
-                    Option::None => "None".to_string(),
-                };
-                let values: Vec<String> = values.iter().map(|v| v.fmt_with_ctx(ctx)).collect();
-                format!("ConstAdt {} [{}]", variant_id, values.join(", "))
-            }
-            RawConstantExpr::Global(id, generics) => {
-                format!(
-                    "{}{}",
-                    ctx.format_object(*id),
-                    generics.fmt_with_ctx_split_trait_refs(ctx)
-                )
-            }
-            RawConstantExpr::TraitConst(trait_ref, name) => {
-                format!("{}::{name}", trait_ref.fmt_with_ctx(ctx),)
-            }
-            RawConstantExpr::Ref(cv) => {
-                format!("&{}", cv.fmt_with_ctx(ctx))
-            }
-            RawConstantExpr::Var(id) => format!("{}", ctx.format_object(*id)),
-            RawConstantExpr::FnPtr(f) => {
-                format!("{}", f.fmt_with_ctx(ctx),)
-            }
-        }
-    }
-}
-
-impl FnPtr {
-    pub fn fmt_with_ctx<C>(&self, ctx: &C) -> String
-    where
-        C: AstFormatter,
-    {
-        let generics = self.generics.fmt_with_ctx_split_trait_refs(ctx);
-        let f = match &self.func {
-            FunIdOrTraitMethodRef::Fun(FunId::Regular(def_id)) => {
-                format!("{}", ctx.format_object(*def_id),)
-            }
-            FunIdOrTraitMethodRef::Fun(FunId::Assumed(assumed)) => {
-                format!("@{}", assumed.variant_name())
-            }
-            FunIdOrTraitMethodRef::Trait(trait_ref, method_id, _) => {
-                format!("{}::{}", trait_ref.fmt_with_ctx(ctx), &method_id.0)
-            }
-        };
-
-        format!("{}{}", f, generics)
-    }
-}
-
-impl std::fmt::Display for ConstantExpr {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::result::Result<(), std::fmt::Error> {
-        write!(f, "{}", self.fmt_with_ctx(&FmtCtx::new()))
-    }
-}
-
-impl Operand {
-    pub fn fmt_with_ctx<C>(&self, ctx: &C) -> String
-    where
-        C: AstFormatter,
-    {
-        match self {
-            Operand::Copy(p) => format!("copy ({})", p.fmt_with_ctx(ctx)),
-            Operand::Move(p) => format!("move ({})", p.fmt_with_ctx(ctx)),
-            Operand::Const(c) => format!("const ({})", c.fmt_with_ctx(ctx)),
-        }
-    }
-}
-
-impl std::fmt::Display for Operand {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::result::Result<(), std::fmt::Error> {
-        write!(f, "{}", self.fmt_with_ctx(&FmtCtx::new()))
-    }
-}
-
-impl Rvalue {
-    pub fn fmt_with_ctx<C>(&self, ctx: &C) -> String
-    where
-        C: AstFormatter,
-    {
-        match self {
-            Rvalue::Use(x) => x.fmt_with_ctx(ctx),
-            Rvalue::Ref(place, borrow_kind) => match borrow_kind {
-                BorrowKind::Shared => format!("&{}", place.fmt_with_ctx(ctx)),
-                BorrowKind::Mut => format!("&mut {}", place.fmt_with_ctx(ctx)),
-                BorrowKind::TwoPhaseMut => {
-                    format!("&two-phase-mut {}", place.fmt_with_ctx(ctx))
-                }
-                BorrowKind::Shallow => format!("&shallow {}", place.fmt_with_ctx(ctx)),
-            },
-            Rvalue::UnaryOp(unop, x) => {
-                format!("{}({})", unop.fmt_with_ctx(ctx), x.fmt_with_ctx(ctx))
-            }
-            Rvalue::BinaryOp(binop, x, y) => {
-                format!("{} {} {}", x.fmt_with_ctx(ctx), binop, y.fmt_with_ctx(ctx))
-            }
-            Rvalue::Discriminant(p, _) => {
-                format!("@discriminant({})", p.fmt_with_ctx(ctx),)
-            }
-            Rvalue::Aggregate(kind, ops) => {
-                let ops_s: Vec<String> = ops.iter().map(|op| op.fmt_with_ctx(ctx)).collect();
-                match kind {
-                    AggregateKind::Adt(def_id, variant_id, _) => {
-                        match def_id {
-                            TypeId::Tuple => format!("({})", ops_s.join(", ")),
-                            TypeId::Assumed(_) => unreachable!(),
-                            TypeId::Adt(def_id) => {
-                                // Format every field
-                                let mut fields = vec![];
-                                for (i, op) in ops.iter().enumerate() {
-                                    let field_id = FieldId::Id::new(i);
-                                    let field_name =
-                                        ctx.format_object((*def_id, *variant_id, field_id));
-                                    fields.push(format!(
-                                        "{}: {}",
-                                        field_name,
-                                        op.fmt_with_ctx(ctx)
-                                    ));
-                                }
-
-                                let variant = match variant_id {
-                                    None => ctx.format_object(*def_id),
-                                    Some(variant_id) => ctx.format_object((*def_id, *variant_id)),
-                                };
-                                format!("{} {{ {} }}", variant, fields.join(", "))
-                            }
-                        }
-                    }
-                    AggregateKind::Array(_, len) => {
-                        format!("[{}; {}]", ops_s.join(", "), len.fmt_with_ctx(ctx))
-                    }
-                    AggregateKind::Closure(fn_id, generics) => {
-                        format!(
-                            "{{{}{}}} {{{}}}",
-                            ctx.format_object(*fn_id),
-                            generics.fmt_with_ctx_split_trait_refs(ctx),
-                            ops_s.join(", ")
-                        )
-                    }
-                }
-            }
-            Rvalue::Global(gid, generics) => {
-                format!(
-                    "{}{}",
-                    ctx.format_object(*gid),
-                    generics.fmt_with_ctx_split_trait_refs(ctx)
-                )
-            }
-            Rvalue::Len(place, ..) => format!("len({})", place.fmt_with_ctx(ctx)),
-            Rvalue::Repeat(op, _ty, cg) => {
-                format!("[{}; {}]", op.fmt_with_ctx(ctx), cg.fmt_with_ctx(ctx))
-            }
-        }
-    }
-}
-
-impl std::fmt::Display for Rvalue {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::result::Result<(), std::fmt::Error> {
-        write!(f, "{}", self.fmt_with_ctx(&FmtCtx::new()))
     }
 }
