+++ charon/src/ast/types_utils.rs
@@ -1,7 +1,4 @@
 //! This file groups everything which is linked to implementations about [crate::types]
-use crate::assumed::get_name_from_type_id;
-use crate::common::TAB_INCR;
-use crate::formatter::{AstFormatter, FmtCtx};
 use crate::types::*;
 use crate::values::*;
 use hax_frontend_exporter as hax;
@@ -25,44 +22,9 @@ impl DeBruijnId {
     }
 }
 
-impl Region {
-    pub fn fmt_with_ctx<C>(&self, ctx: &C) -> String
-    where
-        C: AstFormatter,
-    {
-        match self {
-            Region::Static => "'static".to_string(),
-            Region::BVar(grid, id) => ctx.format_object((*grid, *id)),
-            Region::Erased => "'_".to_string(),
-            Region::Unknown => "'_UNKNOWN_".to_string(),
-        }
-    }
-}
-
 impl TypeVar {
     pub fn new(index: TypeVarId::Id, name: String) -> TypeVar {
         TypeVar { index, name }
-    }
-}
-
-impl std::string::ToString for TypeVar {
-    fn to_string(&self) -> String {
-        self.name.to_string()
-    }
-}
-
-impl std::string::ToString for RegionVar {
-    fn to_string(&self) -> String {
-        match &self.name {
-            Some(name) => name.to_string(),
-            None => format!("'_{}", self.index),
-        }
-    }
-}
-
-impl std::string::ToString for ConstGenericVar {
-    fn to_string(&self) -> String {
-        format!("const {} : {}", self.name, self.ty.to_string())
     }
 }
 
@@ -89,117 +51,6 @@ impl GenericParams {
             trait_clauses: Vec::new(),
         }
     }
-
-    pub fn fmt_with_ctx<C>(&self, ctx: &C) -> String
-    where
-        C: AstFormatter,
-    {
-        if self.is_empty() {
-            "".to_string()
-        } else {
-            let mut params = Vec::new();
-            let GenericParams {
-                regions,
-                types,
-                const_generics,
-                trait_clauses,
-            } = self;
-            for x in regions {
-                params.push(x.to_string());
-            }
-            for x in types {
-                params.push(x.to_string());
-            }
-            for x in const_generics {
-                params.push(x.to_string());
-            }
-            for x in trait_clauses {
-                params.push(x.fmt_with_ctx(ctx));
-            }
-            format!("<{}>", params.join(", "))
-        }
-    }
-
-    pub fn fmt_with_ctx_with_trait_clauses<C>(&self, ctx: &C) -> (String, Vec<String>)
-    where
-        C: AstFormatter,
-    {
-        let mut params = Vec::new();
-        let GenericParams {
-            regions,
-            types,
-            const_generics,
-            trait_clauses,
-        } = self;
-        for x in regions {
-            params.push(x.to_string());
-        }
-        for x in types {
-            params.push(x.to_string());
-        }
-        for x in const_generics {
-            params.push(x.to_string());
-        }
-        let params = if params.is_empty() {
-            "".to_string()
-        } else {
-            format!("<{}>", params.join(", "))
-        };
-
-        let mut clauses = Vec::new();
-        for x in trait_clauses {
-            clauses.push(x.fmt_with_ctx(ctx));
-        }
-        (params, clauses)
-    }
-}
-
-/// [num_parent_clauses]: we store in the definitions all the clauses
-/// they have access to, which includes the clauses inherited from the parent.
-/// This can be confusing: we insert a delimiter between the inherited clauses
-/// and the local clauses.
-pub fn fmt_where_clauses(tab: &str, num_parent_clauses: usize, clauses: Vec<String>) -> String {
-    if clauses.is_empty() {
-        "".to_string()
-    } else {
-        let mut clauses = clauses
-            .iter()
-            .map(|x| format!("\n{tab}{TAB_INCR}{x},"))
-            .collect::<Vec<String>>();
-        if num_parent_clauses > 0 {
-            let local_clauses = clauses.split_off(num_parent_clauses);
-
-            let delim1 = if local_clauses.is_empty() {
-                "".to_string()
-            } else {
-                format!("\n{tab}{TAB_INCR}// Local clauses:")
-            };
-
-            let delim0 = if clauses.is_empty() {
-                "".to_string()
-            } else {
-                format!("\n{tab}{TAB_INCR}// Inherited clauses:")
-            };
-
-            let clauses = clauses.join("");
-            let local_clauses = local_clauses.join("");
-            format!("\n{tab}where{delim0}{clauses}{delim1}{local_clauses}")
-        } else {
-            let clauses = clauses.join("");
-            format!("\n{tab}where{clauses}")
-        }
-    }
-}
-
-impl TraitTypeConstraint {
-    pub fn fmt_with_ctx<C>(&self, ctx: &C) -> String
-    where
-        C: AstFormatter,
-    {
-        let trait_ref = self.trait_ref.fmt_with_ctx(ctx);
-        let ty = self.ty.fmt_with_ctx(ctx);
-        format!("{}::{} = {}", trait_ref, self.type_name, ty)
-    }
 }
 
 impl Predicates {
@@ -210,71 +61,6 @@ impl Predicates {
             trait_type_constraints,
         } = self;
         regions_outlive.is_empty() && types_outlive.is_empty() && trait_type_constraints.is_empty()
-    }
-}
-
-pub fn fmt_where_clauses_with_ctx<C>(
-    ctx: &C,
-    tab: &str,
-    info: &Option<ParamsInfo>,
-    mut trait_clauses: Vec<String>,
-    preds: &Predicates,
-) -> String
-where
-    C: AstFormatter,
-{
-    let mut types_outlive: Vec<_> = preds
-        .types_outlive
-        .iter()
-        .map(|OutlivesPred(x, y)| format!("{} : {}", x.fmt_with_ctx(ctx), y.fmt_with_ctx(ctx)))
-        .collect();
-    let mut regions_outlive: Vec<_> = preds
-        .regions_outlive
-        .iter()
-        .map(|OutlivesPred(x, y)| format!("{} : {}", x.fmt_with_ctx(ctx), y.fmt_with_ctx(ctx)))
-        .collect();
-    let mut type_constraints: Vec<_> = preds
-        .trait_type_constraints
-        .iter()
-        .map(|x| x.fmt_with_ctx(ctx))
-        .collect();
-    match info {
-        None => {
-            let clauses: Vec<_> = trait_clauses
-                .into_iter()
-                .chain(types_outlive.into_iter())
-                .chain(regions_outlive.into_iter())
-                .chain(type_constraints.into_iter())
-                .collect();
-            fmt_where_clauses(tab, 0, clauses)
-        }
-        Some(info) => {
-            // Below: definitely not efficient nor convenient, but it is not really
-            // important
-            let local_clauses: Vec<_> = trait_clauses
-                .split_off(info.num_trait_clauses)
-                .into_iter()
-                .chain(regions_outlive.split_off(info.num_regions_outlive))
-                .chain(types_outlive.split_off(info.num_types_outlive).into_iter())
-                .chain(
-                    type_constraints
-                        .split_off(info.num_trait_type_constraints)
-                        .into_iter(),
-                )
-                .collect();
-            let inherited_clauses: Vec<_> = trait_clauses
-                .into_iter()
-                .chain(regions_outlive.into_iter())
-                .chain(types_outlive.into_iter())
-                .chain(type_constraints.into_iter())
-                .collect();
-            let num_inherited = inherited_clauses.len();
-            let all_clauses: Vec<_> = inherited_clauses
-                .into_iter()
-                .chain(local_clauses.into_iter())
-                .collect();
-            fmt_where_clauses(tab, num_inherited, all_clauses)
-        }
     }
 }
 
@@ -349,163 +135,6 @@ impl GenericArgs {
             },
         )
     }
-
-    pub(crate) fn fmt_with_ctx_no_brackets<C>(&self, ctx: &C) -> String
-    where
-        C: AstFormatter,
-    {
-        let mut params = Vec::new();
-        let GenericArgs {
-            regions,
-            types,
-            const_generics,
-            trait_refs,
-        } = self;
-        for x in regions {
-            params.push(x.fmt_with_ctx(ctx));
-        }
-        for x in types {
-            params.push(x.fmt_with_ctx(ctx));
-        }
-        for x in const_generics {
-            params.push(x.fmt_with_ctx(ctx));
-        }
-        for x in trait_refs {
-            params.push(x.fmt_with_ctx(ctx))
-        }
-        params.join(", ")
-    }
-
-    pub fn fmt_with_ctx<C>(&self, ctx: &C) -> String
-    where
-        C: AstFormatter,
-    {
-        if self.is_empty() {
-            "".to_string()
-        } else {
-            format!("<{}>", self.fmt_with_ctx_no_brackets(ctx),)
-        }
-    }
-
-    pub fn fmt_with_ctx_split_trait_refs<C>(&self, ctx: &C) -> String
-    where
-        C: AstFormatter,
-    {
-        let mut params = Vec::new();
-        let GenericArgs {
-            regions,
-            types,
-            const_generics,
-            trait_refs,
-        } = self;
-        for x in regions {
-            params.push(x.fmt_with_ctx(ctx));
-        }
-        for x in types {
-            params.push(x.fmt_with_ctx(ctx));
-        }
-        for x in const_generics {
-            params.push(x.fmt_with_ctx(ctx));
-        }
-        let params = if params.is_empty() {
-            "".to_string()
-        } else {
-            format!("<{}>", params.join(", "))
-        };
-
-        let mut clauses = Vec::new();
-        for x in trait_refs {
-            clauses.push(x.fmt_with_ctx(ctx));
-        }
-        let clauses = if clauses.is_empty() {
-            "".to_string()
-        } else {
-            format!("[{}]", clauses.join(", "))
-        };
-        format!("{params}{clauses}")
-    }
-}
-
-impl TraitClause {
-    pub fn fmt_with_ctx<C>(&self, ctx: &C) -> String
-    where
-        C: AstFormatter,
-    {
-        let clause_id = ctx.format_object(self.clause_id);
-        let trait_id = ctx.format_object(self.trait_id);
-        let generics = self.generics.fmt_with_ctx(ctx);
-        format!("[{clause_id}]: {trait_id}{generics}")
-    }
-}
-
-impl TraitInstanceId {
-    pub fn fmt_with_ctx<C>(&self, ctx: &C) -> String
-    where
-        C: AstFormatter,
-    {
-        match self {
-            TraitInstanceId::SelfId => "Self".to_string(),
-            TraitInstanceId::ParentClause(id, _decl_id, clause_id) => {
-                let id = id.fmt_with_ctx(ctx);
-                // Using on purpose [to_pretty_string] instead of [format_object]:
-                // the clause is local to the associated type, so it should not
-                // be referenced in the current context.
-                let clause = clause_id.to_pretty_string();
-                format!("(parents({id})::[{clause}])")
-            }
-            TraitInstanceId::ItemClause(id, _decl_id, type_name, clause_id) => {
-                let id = id.fmt_with_ctx(ctx);
-                // Using on purpose [to_pretty_string] instead of [format_object]:
-                // the clause is local to the associated type, so it should not
-                // be referenced in the current context.
-                let clause = clause_id.to_pretty_string();
-                format!("({id}::{type_name}::[{clause}])")
-            }
-            TraitInstanceId::TraitImpl(id) => ctx.format_object(*id),
-            TraitInstanceId::Clause(id) => ctx.format_object(*id),
-            TraitInstanceId::BuiltinOrAuto(id) => ctx.format_object(*id),
-            TraitInstanceId::FnPointer(box ty) => {
-                format!("(fn_ptr:{})", ty.fmt_with_ctx(ctx))
-            }
-            TraitInstanceId::Closure(fid, generics) => {
-                format!(
-                    "(closure:{}{})",
-                    ctx.format_object(*fid),
-                    generics.fmt_with_ctx(ctx)
-                )
-            }
-            TraitInstanceId::Unsolved(trait_id, generics) => {
-                format!(
-                    "Unsolved({}{})",
-                    ctx.format_object(*trait_id),
-                    generics.fmt_with_ctx(ctx),
-                )
-            }
-            TraitInstanceId::Unknown(msg) => format!("UNKNOWN({msg})"),
-        }
-    }
-}
-
-impl TraitRef {
-    pub fn fmt_with_ctx<C>(&self, ctx: &C) -> String
-    where
-        C: AstFormatter,
-    {
-        let trait_id = self.trait_id.fmt_with_ctx(ctx);
-        let generics = self.generics.fmt_with_ctx_split_trait_refs(ctx);
-        format!("{trait_id}{generics}")
-    }
-}
-
-impl TraitDeclRef {
-    pub fn fmt_with_ctx<C>(&self, ctx: &C) -> String
-    where
-        C: AstFormatter,
-    {
-        let trait_id = ctx.format_object(self.trait_id);
-        let generics = self.generics.fmt_with_ctx_split_trait_refs(ctx);
-        format!("{trait_id}{generics}")
-    }
 }
 
 impl TypeDecl {
@@ -527,98 +156,6 @@ impl TypeDecl {
             }
             TypeDeclKind::Error(_) => Err(()),
         }
-    }
-
-    pub fn fmt_with_ctx<C>(&self, ctx: &C) -> String
-    where
-        C: AstFormatter,
-    {
-        let ctx = &ctx.set_generics(&self.generics);
-
-        let (params, trait_clauses) = self.generics.fmt_with_ctx_with_trait_clauses(ctx);
-        // Predicates
-        let eq_space = if trait_clauses.is_empty() && self.preds.is_empty() {
-            " ".to_string()
-        } else {
-            "\n ".to_string()
-        };
-        let preds = fmt_where_clauses_with_ctx(ctx, "  ", &None, trait_clauses, &self.preds);
-
-        match &self.kind {
-            TypeDeclKind::Struct(fields) => {
-                if !fields.is_empty() {
-                    let fields: Vec<String> = fields
-                        .iter()
-                        .map(|f| format!("\n  {}", f.fmt_with_ctx(ctx)))
-                        .collect();
-                    let fields = fields.join(",");
-                    format!(
-                        "struct {}{params}{preds}{eq_space}=\n{{{fields}\n}}",
-                        self.name.fmt_with_ctx(ctx)
-                    )
-                } else {
-                    format!(
-                        "struct {}{params}{preds}{eq_space}= {{}}",
-                        self.name.fmt_with_ctx(ctx)
-                    )
-                }
-            }
-            TypeDeclKind::Enum(variants) => {
-                let variants: Vec<String> = variants
-                    .iter()
-                    .map(|v| format!("|  {}", v.fmt_with_ctx(ctx)))
-                    .collect();
-                let variants = variants.join("\n");
-                format!(
-                    "enum {}{params}{preds}{eq_space}=\n{variants}\n",
-                    self.name.fmt_with_ctx(ctx)
-                )
-            }
-            TypeDeclKind::Opaque => {
-                format!("opaque type {}{params}{preds}", self.name.fmt_with_ctx(ctx))
-            }
-            TypeDeclKind::Error(msg) => {
-                format!(
-                    "opaque type {}{params}{preds} = ERROR({msg})",
-                    self.name.fmt_with_ctx(ctx),
-                )
-            }
-        }
-    }
-}
-
-impl Variant {
-    pub fn fmt_with_ctx<C>(&self, ctx: &C) -> String
-    where
-        C: AstFormatter,
-    {
-        let fields: Vec<String> = self.fields.iter().map(|f| f.fmt_with_ctx(ctx)).collect();
-        let fields = fields.join(", ");
-        format!("{}({})", self.name, fields)
-    }
-}
-
-impl Field {
-    pub fn fmt_with_ctx<C>(&self, ctx: &C) -> String
-    where
-        C: AstFormatter,
-    {
-        match &self.name {
-            Option::Some(name) => format!("{}: {}", name, self.ty.fmt_with_ctx(ctx)),
-            Option::None => self.ty.fmt_with_ctx(ctx),
-        }
-    }
-}
-
-impl std::string::ToString for Variant {
-    fn to_string(&self) -> String {
-        self.fmt_with_ctx(&FmtCtx::new())
-    }
-}
-
-impl std::string::ToString for Field {
-    fn to_string(&self) -> String {
-        self.fmt_with_ctx(&FmtCtx::new())
     }
 }
 
@@ -683,143 +220,8 @@ impl IntegerTy {
     }
 }
 
-impl TypeVarId::Id {
-    pub fn to_pretty_string(self) -> String {
-        format!("@T{self}")
-    }
-}
-
-impl TypeDeclId::Id {
-    pub fn to_pretty_string(self) -> String {
-        format!("@Adt{self}")
-    }
-}
-
-impl VariantId::Id {
-    pub fn to_pretty_string(self) -> String {
-        format!("@Variant{self}")
-    }
-}
-
-impl FieldId::Id {
-    pub fn to_pretty_string(self) -> String {
-        format!("@Field{self}")
-    }
-}
-
-impl ConstGenericVarId::Id {
-    pub fn to_pretty_string(self) -> String {
-        format!("@Const{self}")
-    }
-}
-
-impl GlobalDeclId::Id {
-    pub fn to_pretty_string(self) -> String {
-        format!("@Global{self}")
-    }
-}
-
-impl TraitClauseId::Id {
-    pub fn to_pretty_string(self) -> String {
-        format!("@TraitClause{self}")
-    }
-}
-
-impl TraitDeclId::Id {
-    pub fn to_pretty_string(self) -> String {
-        format!("@TraitDecl{self}")
-    }
-}
-
-impl TraitImplId::Id {
-    pub fn to_pretty_string(self) -> String {
-        format!("@TraitImpl{self}")
-    }
-}
-
 pub fn bound_region_var_to_pretty_string(grid: DeBruijnId, rid: RegionId::Id) -> String {
     format!("'_{}_{}", grid.index, rid.to_string())
-}
-
-impl std::fmt::Display for IntegerTy {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::result::Result<(), std::fmt::Error> {
-        match self {
-            IntegerTy::Isize => write!(f, "isize"),
-            IntegerTy::I8 => write!(f, "i8"),
-            IntegerTy::I16 => write!(f, "i16"),
-            IntegerTy::I32 => write!(f, "i32"),
-            IntegerTy::I64 => write!(f, "i64"),
-            IntegerTy::I128 => write!(f, "i128"),
-            IntegerTy::Usize => write!(f, "usize"),
-            IntegerTy::U8 => write!(f, "u8"),
-            IntegerTy::U16 => write!(f, "u16"),
-            IntegerTy::U32 => write!(f, "u32"),
-            IntegerTy::U64 => write!(f, "u64"),
-            IntegerTy::U128 => write!(f, "u128"),
-        }
-    }
-}
-
-impl std::fmt::Display for LiteralTy {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::result::Result<(), std::fmt::Error> {
-        match self {
-            LiteralTy::Integer(ty) => ty.fmt(f),
-            LiteralTy::Char => write!(f, "char"),
-            LiteralTy::Bool => write!(f, "bool"),
-        }
-    }
-}
-
-// IntTy is not defined in the current crate
-pub fn intty_to_string(ty: hax::IntTy) -> String {
-    use hax::IntTy::*;
-    match ty {
-        Isize => "isize".to_string(),
-        I8 => "i8".to_string(),
-        I16 => "i16".to_string(),
-        I32 => "i32".to_string(),
-        I64 => "i64".to_string(),
-        I128 => "i128".to_string(),
-    }
-}
-
-// UintTy is not defined in the current crate
-pub fn uintty_to_string(ty: hax::UintTy) -> String {
-    use hax::UintTy::*;
-    match ty {
-        Usize => "usize".to_string(),
-        U8 => "u8".to_string(),
-        U16 => "u16".to_string(),
-        U32 => "u32".to_string(),
-        U64 => "u64".to_string(),
-        U128 => "u128".to_string(),
-    }
-}
-
-impl TypeId {
-    pub fn fmt_with_ctx<C>(&self, ctx: &C) -> String
-    where
-        C: AstFormatter,
-    {
-        match self {
-            TypeId::Tuple => "".to_string(),
-            TypeId::Adt(def_id) => ctx.format_object(*def_id),
-            TypeId::Assumed(aty) => get_name_from_type_id(*aty).join("::"),
-        }
-    }
-}
-
-impl ConstGeneric {
-    pub fn fmt_with_ctx<C>(&self, ctx: &C) -> String
-    where
-        C: AstFormatter,
-    {
-        match self {
-            ConstGeneric::Var(id) => ctx.format_object(*id),
-            ConstGeneric::Value(v) => v.to_string(),
-            ConstGeneric::Global(id) => ctx.format_object(*id),
-        }
-    }
 }
 
 impl Ty {
@@ -862,77 +264,6 @@ impl Ty {
         }
     }
 
-    /// Format the type as a string.
-    ///
-    /// We take an optional type context to be able to implement the Display
-    /// trait, in which case there is no type context available and we print
-    /// the ADT ids rather than their names.
-    pub fn fmt_with_ctx<C>(&self, ctx: &C) -> String
-    where
-        C: AstFormatter,
-    {
-        match self {
-            Ty::Adt(id, generics) => {
-                let adt_ident = id.fmt_with_ctx(ctx);
-
-                if id.is_tuple() {
-                    assert!(generics.trait_refs.is_empty());
-                    let generics = generics.fmt_with_ctx_no_brackets(ctx);
-                    format!("({generics})")
-                } else {
-                    let generics = generics.fmt_with_ctx(ctx);
-                    format!("{adt_ident}{generics}")
-                }
-            }
-            Ty::TypeVar(id) => ctx.format_object(*id),
-            Ty::Literal(kind) => kind.to_string(),
-            Ty::Never => "!".to_string(),
-            Ty::Ref(r, ty, kind) => match kind {
-                RefKind::Mut => {
-                    format!("&{} mut ({})", r.fmt_with_ctx(ctx), ty.fmt_with_ctx(ctx))
-                }
-                RefKind::Shared => {
-                    format!("&{} ({})", r.fmt_with_ctx(ctx), ty.fmt_with_ctx(ctx))
-                }
-            },
-            Ty::RawPtr(ty, kind) => match kind {
-                RefKind::Mut => format!("*const {}", ty.fmt_with_ctx(ctx)),
-                RefKind::Shared => format!("*mut {}", ty.fmt_with_ctx(ctx)),
-            },
-            Ty::TraitType(trait_ref, name) => {
-                format!("{}::{name}", trait_ref.fmt_with_ctx(ctx),)
-            }
-            Ty::Arrow(regions, inputs, box output) => {
-                // Update the bound regions
-                let ctx = &ctx.push_bound_regions(regions);
-
-                let regions = if regions.is_empty() {
-                    "".to_string()
-                } else {
-                    format!(
-                        "<{}>",
-                        regions
-                            .iter()
-                            .map(|r| r.to_string())
-                            .collect::<Vec<String>>()
-                            .join(", ")
-                    )
-                };
-                let inputs = inputs
-                    .iter()
-                    .map(|x| x.fmt_with_ctx(ctx))
-                    .collect::<Vec<String>>()
-                    .join(", ");
-                if output.is_unit() {
-                    format!("fn{regions}({inputs})")
-                } else {
-                    let output = output.fmt_with_ctx(ctx);
-                    format!("fn{regions}({inputs}) -> {output}")
-                }
-            }
-        }
-    }
-
     /// Return true if the type is Box
     pub fn is_box(&self) -> bool {
         match self {
@@ -955,17 +286,6 @@ impl Ty {
                 Some(generics.types.get(0).unwrap())
             }
             _ => None,
-        }
-    }
-}
-
-impl std::fmt::Display for Region {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::result::Result<(), std::fmt::Error> {
-        match self {
-            Region::Static => write!(f, "'static"),
-            Region::BVar(grid, id) => write!(f, "'_{}_{id}", grid.index),
-            Region::Erased => write!(f, "'_"),
-            Region::Unknown => write!(f, "'_UNKNOWN_"),
         }
     }
 }
@@ -1473,49 +793,3 @@ pub trait TypeVisitor {
 }
 
 } // make_generic_in_borrows
-
-impl FunSig {
-    pub fn fmt_with_ctx<C>(&self, ctx: &C) -> String
-    where
-        C: AstFormatter,
-    {
-        let ctx = &ctx.set_generics(&self.generics);
-
-        // Unsafe keyword
-        let unsafe_kw = if self.is_unsafe {
-            "unsafe ".to_string()
-        } else {
-            "".to_string()
-        };
-
-        // Generic parameters
-        let (params, trait_clauses) = self.generics.fmt_with_ctx_with_trait_clauses(ctx);
-
-        // Arguments
-        let mut args: Vec<String> = Vec::new();
-        for ty in &self.inputs {
-            args.push(ty.fmt_with_ctx(ctx).to_string());
-        }
-        let args = args.join(", ");
-
-        // Return type
-        let ret_ty = &self.output;
-        let ret_ty = if ret_ty.is_unit() {
-            "".to_string()
-        } else {
-            format!(" -> {}", ret_ty.fmt_with_ctx(ctx))
-        };
-
-        // Clauses
-        let clauses = fmt_where_clauses_with_ctx(
-            ctx,
-            "",
-            &self.parent_params_info,
-            trait_clauses,
-            &self.preds,
-        );
-
-        // Put everything together
-        format!("{unsafe_kw}fn{params}({args}){ret_ty}{clauses}",)
-    }
-}
