+++ charon/src/ast/ullbc_ast_utils.rs
@@ -1,7 +1,4 @@
 //! Implementations for [crate::ullbc_ast]
-use crate::common::TAB_INCR;
-use crate::formatter::AstFormatter;
-pub use crate::gast_utils::*;
 use crate::meta::Meta;
 use crate::types::*;
 use crate::ullbc_ast::*;
@@ -36,145 +33,6 @@ impl Statement {
 impl Terminator {
     pub fn new(meta: Meta, content: RawTerminator) -> Self {
         Terminator { meta, content }
-    }
-}
-
-impl Statement {
-    pub fn fmt_with_ctx<C>(&self, ctx: &C) -> String
-    where
-        C: AstFormatter,
-    {
-        match &self.content {
-            RawStatement::Assign(place, rvalue) => format!(
-                "{} := {}",
-                place.fmt_with_ctx(ctx),
-                rvalue.fmt_with_ctx(ctx),
-            ),
-            RawStatement::FakeRead(place) => {
-                format!("@fake_read({})", place.fmt_with_ctx(ctx))
-            }
-            RawStatement::SetDiscriminant(place, variant_id) => format!(
-                "@discriminant({}) := {}",
-                place.fmt_with_ctx(ctx),
-                variant_id
-            ),
-            RawStatement::StorageDead(vid) => {
-                format!("@storage_dead({})", vid.to_pretty_string())
-            }
-            RawStatement::Deinit(place) => {
-                format!("@deinit({})", place.fmt_with_ctx(ctx))
-            }
-        }
-    }
-}
-
-impl Terminator {
-    pub fn fmt_with_ctx<C>(&self, ctx: &C) -> String
-    where
-        C: AstFormatter,
-    {
-        match &self.content {
-            RawTerminator::Goto { target } => format!("goto bb{target}"),
-            RawTerminator::Switch { discr, targets } => match targets {
-                SwitchTargets::If(true_block, false_block) => format!(
-                    "if {} -> bb{} else -> bb{}",
-                    discr.fmt_with_ctx(ctx),
-                    true_block,
-                    false_block
-                ),
-                SwitchTargets::SwitchInt(_ty, maps, otherwise) => {
-                    let mut maps: Vec<String> = maps
-                        .iter()
-                        .map(|(v, bid)| format!("{}: bb{}", v.to_string(), bid))
-                        .collect();
-                    maps.push(format!("otherwise: bb{otherwise}"));
-                    let maps = maps.join(", ");
-
-                    format!("switch {} -> {}", discr.fmt_with_ctx(ctx), maps)
-                }
-            },
-            RawTerminator::Panic => "panic".to_string(),
-            RawTerminator::Return => "return".to_string(),
-            RawTerminator::Unreachable => "unreachable".to_string(),
-            RawTerminator::Drop { place, target } => {
-                format!("drop {} -> bb{}", place.fmt_with_ctx(ctx), target)
-            }
-            RawTerminator::Call { call, target } => {
-                let (call_s, _) = fmt_call(ctx, call);
-                format!("{} := {call_s} -> bb{target}", call.dest.fmt_with_ctx(ctx),)
-            }
-            RawTerminator::Assert {
-                cond,
-                expected,
-                target,
-            } => format!(
-                "assert({} == {}) -> bb{}",
-                cond.fmt_with_ctx(ctx),
-                expected,
-                target
-            ),
-        }
-    }
-}
-
-impl BlockData {
-    pub fn fmt_with_ctx<C>(&self, tab: &str, ctx: &C) -> String
-    where
-        C: AstFormatter,
-    {
-        let mut out: Vec<String> = Vec::new();
-
-        // Format the statements
-        for statement in &self.statements {
-            out.push(format!("{}{};\n", tab, statement.fmt_with_ctx(ctx)).to_string());
-        }
-
-        // Format the terminator
-        out.push(format!("{}{};", tab, self.terminator.fmt_with_ctx(ctx)));
-
-        // Join the strings
-        out.join("")
-    }
-}
-
-pub(crate) fn fmt_body_blocks_with_ctx<C>(
-    body: &BlockId::Vector<BlockData>,
-    tab: &str,
-    ctx: &C,
-) -> String
-where
-    C: AstFormatter,
-{
-    let block_tab = format!("{tab}{TAB_INCR}");
-    let mut blocks: Vec<String> = Vec::new();
-    for (bid, block) in body.iter_indexed_values() {
-        blocks.push(
-            format!(
-                "{tab}bb{}: {{\n{}\n{tab}}}\n",
-                bid.index(),
-                block.fmt_with_ctx(&block_tab, ctx),
-            )
-            .to_string(),
-        );
-    }
-    blocks.join("\n")
-}
-
-impl FunDecl {
-    pub fn fmt_with_ctx<C>(&self, ctx: &C) -> String
-    where
-        C: AstFormatter,
-    {
-        self.gfmt_with_ctx("", ctx)
-    }
-}
-
-impl GlobalDecl {
-    pub fn fmt_with_ctx<C>(&self, ctx: &C) -> String
-    where
-        C: AstFormatter,
-    {
-        self.gfmt_with_ctx("", ctx)
     }
 }
