+++ charon/src/pretty/mod.rs
@@ -1,0 +1,3 @@
+//! This module contains functions to pretty-print charon types.
+pub mod fmt_with_ctx;
+pub mod formatter;
